requirements.txt
asgiref==3.11.0
certifi==2025.11.12
distlib==0.4.0
dj-rest-auth==7.1.1
Django==5.2
django-allauth==65.14.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.15
filelock==3.20.1
gunicorn==24.1.1
inflection==0.5.1
packaging==25.0
pipenv==2026.0.3
platformdirs==4.5.1
PyJWT==2.11.0
pytz==2026.1.post1
PyYAML==6.0.3
setuptools==80.9.0
sqlparse==0.5.5
uritemplate==4.2.0
virtualenv==20.35.4
python-dotenv